Swap name of QPro and QProtocol
