+++ tests/classes/baseOthers/test_QSimCom.py
@@ -1,7 +1,7 @@
 import pytest
 import quanguru as qg
 
-@pytest.mark.parametrize('clas', [qg.QuantumSystem, qg.QPro.genericProtocol])
+@pytest.mark.parametrize('clas', [qg.QuantumSystem, qg.QProtocol.genericProtocol])
 def test_setAndGetInternalSimulationParametersThroughSelf(clas):
     inst1 = clas(stepSize=1)
     assert inst1.stepSize == inst1.simulation.stepSize
